upgrade string/locale cms collection to utility functions
