+++ example/content/locales/en/translations.json
@@ -1,7 +1,5 @@
 {
   "Hello": "Hello",
   "404.title": "404: Not found",
-  "404.message": "You just hit a route that doesn't exist... sadness doesn't exist only suffering does",
-  "posts.title": "Blog",
-  "tags.title": "Tags"
+  "404.message": "You just hit a route that doesn't exist... sadness doesn't exist only suffering does"
 }
