+++ gatsby-theme-w3f/src/cms/collections/index.js
@@ -1,5 +1,14 @@
-import strings from './strings';
+import translations from '../../../content/locales/en/translations.json';
+import {buildLocalesCollection} from './locales';
+
+const stringFiles = [
+  {
+    name: 'translations',
+    strings: translations,
+  }
+]
+const locales = buildLocalesCollection(stringFiles);
 
 export default {
-  strings
+  locales
 };
